fix(core): Fork scope if custom scope is passed to `startSpan` (#14900)
Fix unexpected behaviour that would emerge when starting
multiple spans in sequence with `startSpan` when passing on the same
scope. Prior to this change, the second span would be started as the
child span of the first one, because the span set active on the custom
scope was not re-set to the parent span of the first span. 

This patch fixes this edge case by also forking the custom scope if it is
passed into `startSpan`.