fix(node): Enforce that ContextLines integration does not leave open file handles (#14995)
The ContextLines integration uses readable streams to more memory
efficiently read files that it uses to attach source context to outgoing
events. See more details here:
https://github.com/getsentry/sentry-javascript/pull/12221

Unfortunately, if we don't explicitly destroy the stream after creating
and using it, it won't get closed, even when we remove the readline
interface that uses the stream (which actual does the reading of files).

To fix this, we adjust the resolve logic when getting file context to
destroy the stream, as we anyway are done with the readline interface.