feat(react)!: Update `ErrorBoundary` `componentStack` type (#14742)
The principle thing that drove this change was this todo:

https://github.com/getsentry/sentry-javascript/blob/5b773779693cb52c9173c67c42cf2a9e48e927cb/packages/react/src/errorboundary.tsx#L101-L102

Specifically we wanted to remove `null` as a valid state from
`componentStack`, making sure that all exposed public API see it as
`string | undefined`. React always returns a `string` `componentStack`
from the error boundary, so this matches our typings more closely to
react.

By making this change, we can also clean up the `render` logic a little.
Specifically we can check for `state.componentStack === null` to
determine if a fallback is rendered, and then I went ahead and removed
some unnecessary nesting.