feat(core)!: Stop accepting `event` as argument for `recordDroppedEvent` (#14999)
The event was no longer used for some time, instead you can (optionally)
pass a count of dropped events as third argument. This has been around
since v7, so we can finally clean this up here, which should also save
some bytes.

This was not really deprecated before, but is also more an intimate API
so I do not expect this to affect users too much. Also, deprecating
arguments like this does not really work well with eslint anyhow, so
even if we would deprecate it it is unlikely users would find out
anyhow.