chore(v8/deps): Upgrade to Vitest 2.1.8 and Vite 5.4.11 (#15038)
Backport of #14971 to v8. This is a prerequisite for an eventual
backport of #14968, which will itself will make applying/backporting
#14967 much easier.

I ran `yarn build` and `yarn test` locally. Building succeeded, and all
but 2 test suites passed. The two that failed also failed for me on `v8`
(without any of my changes), so I'm assuming it's something to do with
my environment.