feat(core): Add client outcomes for breadcrumbs buffer (#15082)
ref https://github.com/getsentry/team-sdks/issues/116

This PR implements a new client discard reason for `buffer_overflow`.
This will be used to track when the internal breadcrumbs buffer
overflows for the new logs product that we are working on. This is
documented in develop here:
https://github.com/getsentry/sentry-docs/pull/12395

Note: The reason we have `buffer_overflow` as a separate item to
`queue_overflow` is that in the future when we send log items in
envelopes we'll increment `queue_overflow` for the transport queue. We
want to differentiate between the transport queue and the internal
buffer explicitly.