ref(core)!: Move `shutdownTimeout` option type from core to node (#15217)
We used to have `shutdownTimeout` in the core client options, but this
option only applies in node, so we can move it there. Also fixing a
small unnecessary check I noticed along the way.

This is technically breaking, but should not affect anybody as this
option is not used (and documented) anywhere outside of node.