feat(core): Allow to configure `disableIntegrations` to opt-out
This new top-level config allows to opt-out of any added integration, ensuring it is not actually added.

This is mainly designed to opt-out of default integrations, but will also apply to any manually added integration.

There are type hints that should help with usage there, but any key is allowed to keep this flexible. Type hints are manually configured (could not find a way to infer this from integrations, as the names are not statically exposed there...) for now.

Usage:

```js
Sentry.init({
  disableIntegrations: { BrowserSession: true, InboundFilters: false }
});
```

Will disable the browserSessionIntegration but not the inboundFilters one.