feat(core): Rename `hasTracingEnabled` to `hasSpansEnabled` (#15309)
Rename the `hasTracingEnabled` utility function to
`hasSpansEnabled`. The original name is causing confusion and doesn't
represent what the function actually does: determining if the SDK is
configured to record spans. This function does not check if trace
propagation without span recording (or "Tracing without Performance") is
enabled (which it basically always is).