ref(svelte): Remove SvelteKit detection (#15313)
When we first created the Svelte SDK, we wanted to measure how often it
was used in a SvelteKit application. Therefore we added some SvelteKit
detection logic to annotate events coming from the Svelte SDK. Since we
have a SvelteKit SDK now, we don't need to track this anymore. Saves a
few bytes, too.

This patch therefore removes the detection and event processing logic