fix(nuxt): Only use filename with file extension from command (#15445)
The Nuxt SDK also uses the preview command to determine where to put the
Sentry top-level import. Like this: `node .output/server/index.mjs`
(would add Sentry to `index.mjs`)

However, the function to get the file name also used folder names
without file extensions which led to Sentry not being included at the
top of the index file.

This was a problem in e.g. Azure as the command is the following: `npx
@azure/static-web-apps-cli start ./public --api-location ./server`