feat(replay): Expose rrweb recordCrossOriginIframes under _experiments (#14916)
Closes #14809.

I have tested this with my project and it sort of works. While it's not
perfect (sometimes produces an unplayable replay), it might unblock
other users looking to experiment with this feature.

Verify:
- [ ] If you've added code that should be tested, please add tests.
- [x] Ensure your code lints and the test suite passes (`yarn lint`) &
(`yarn test`).

This just exposes the API under experiments, as such I don't think it
requires tests.

---------

Co-authored-by: Billy Vong <billyvg@users.noreply.github.com>