feat(node): Support Express v5 (#15380)
The goal is for Express to eventually support publishing events to
`diagnostics_channel` (#15107) so that Open Telemetry can instrument it
without monkey-patching internal code. However, this might take a while
and it would be great to support Express v5 now. This PR is a stop-gap
solution until that work is complete and published.

This PR vendors the code added in my otel PR: 
- https://github.com/open-telemetry/opentelemetry-js-contrib/pull/2437 
- Adds a new instrumentation specifically for hooking express v5
- Copies the Express v4 integration tests to test v5
- The only changes in the tests is the removal of a couple of complex
regex tests where the regexes are no longer supported by Express.
- Modifies the NestJs v11 tests which now support full Express spans