ref(flags/v8): rename unleash integration param (#15345)
Renaming this parameter to keep the name consistent with
https://github.com/getsentry/sentry-javascript/pull/15319. We can reuse
this generic name for future FF integrations.

The old name is still supported but now marked as deprecated. As a
result both params have to be optional, the constructor throws if
neither is given. The old name will not be supported in
https://github.com/getsentry/sentry-javascript/pull/15343.