test(gatsby): Migrate to vitest (#15494)
Before: `6.437 s`

After: `1.23s`

The first thing we did was migrate the `pretest.ts` script to use a
vitest global setup instead. This is cleaner, and should ensure it never
gets lost in refactors either.

After that we did the conversion. The trickiest thing to change was
`packages/gatsby/test/gatsby-node.test.ts`, as `gatsby-node.js` is a CJS
file. This is tricky to patch properly with vitest for the spy. To get
around this, we monkeypatch `Module._load` temporarily to load a stub
which we defined in `vi.hoisted` (which hoists it to the top so it's
always available).