fix(nuxt): Ignore 300-400 status codes on app errors in Nuxt (#15473)
Before submitting a pull request, please take a look at our

[Contributing](https://github.com/getsentry/sentry-javascript/blob/master/CONTRIBUTING.md)
guidelines and verify:

- [ ] If you've added code that should be tested, please add tests.
- [ ] Ensure your code lints and the test suite passes (`yarn lint`) &
(`yarn test`).

### Description

I noticed that Nuxt 300-400 status codes from the app are being reported
to Sentry unnecessarily. There appears to be a guard against doing this
on the server side errors but client side errors like 404 or 401s are
being reported to Sentry.

I have simply added the same guard on the client side error reporter for
the Nuxt integration of Sentry.