feat(core): Default filter unactionable error (#15527)
We've been filtering a specific "undefined is not an object" error that
referenced "a.L" in our error filters. However, similar errors occured
with other uppercase letters (e.g. `a.K`) - they all represent the same
class of non-actionable error.

This pr just updates the filter to use a regex for this pattern.

ref https://github.com/getsentry/sentry/issues/85531