fix(node): Allow for `undefined` transport to be passed in (#15560)
fixes https://github.com/getsentry/sentry-javascript/issues/15558

The Node SDK init options allow for `transport` to be set as
`undefined`, but this breaks because of an error `TypeError:
options.transport is not a function`.

You can see a reproduction of this here:
https://stackblitz.com/edit/stackblitz-starters-phh3lfmy?file=index.ts

To fix this, we make sure we use `makeNodeTransport` if
`options.transport` is set to `undefined`.