Add cloudflare adapter detection and path generation (#15603)
Before submitting a pull request, please take a look at our

[Contributing](https://github.com/getsentry/sentry-javascript/blob/master/CONTRIBUTING.md)
guidelines and verify:

- [x] If you've added code that should be tested, please add tests.
- [x] Ensure your code lints and the test suite passes (`yarn lint`) &
(`yarn test`).

This PR expands upon #14672 by adding detection of the
@sveltejs/adapter-cloudflare adapter and defining it's output directory
(cloudflare as opposed to output). I've tested it with a Cloudflare
Pages project and it was successful in building and uploading
sourcemaps, whereas relying on 'other' caused an infinite loop and an
OOM when trying to upload sourcemaps.