fix(node): Ensure incoming traces are propagated without HttpInstrumentation (#15732)
Part of https://github.com/getsentry/sentry-javascript/pull/15730

Noticed that this is not really happening otherwise, so if we are not
adding the `HttpInstrumentation` we now make sure to extract incoming
traces ourselves! This will unlock us not adding that instrumentation
when tracing is disabled.

Also includes https://github.com/getsentry/sentry-javascript/pull/15731
as a side effect.