feat(core): Add `captureLog` method (#15717)
## Description

`captureLog` takes in a `log` interface as well as a scope, and
constructs a serialized log to send to Sentry. Instead of directly
sending the serialized log, it adds it to a log buffer that is flushed
based on some strategy. As a fallback, the buffer flushes itself when it
gets over 100 items, which ensures we never get a payload thats too
large.

The browser and server runtime client are expected to add more advanced
flushing strategies on top of the simple implementation exposed in
`captureLog`. This will be done in future PRs.

The serialized log that is constructed by `captureLog` send logs via the
opentelemetry format. This is temporary, in the future we'll switch to
the sentry logs protocol, but while that gets ready we are using the
OTEL one.

## Next steps

1. Add flushing strategy for server-runtime-client
2. Implement top-level API methods as per develop spec
3. Add support for paramaterized strings
4. Add integration tests to validate envelope gets sent correctly