fix(nuxt): Delete Nuxt server template injection (#15749)
Follow-up for https://github.com/getsentry/sentry-javascript/pull/15710

The server config was still processed by Nuxt (not only Nitro) because
the file was imported in a plugin template. This code can be deleted
now, as the Sentry server config should only be processed on the
Nitro-side.


fixes https://github.com/getsentry/sentry-javascript/issues/15628