feat(react): Export captureReactException method (#15746)
This PR exports the `captureReactException` method from the React
package, making it available as a public API. Additionally, it refactors
the implementation to use `withScope` instead of directly passing
context to `captureException`. This was done so that the arguments to
`captureReactException` better align to `captureException`.

Users can now directly access and use this method in their applications:

```javascript
import * as Sentry from '@sentry/react';

class ErrorBoundary extends React.Component {
  componentDidCatch(error, info) {
    Sentry.captureReactException(error, info);
  }

  render() {
    return this.props.children;
  }
}
```

The reason why `captureReactException` is named as such and not
`captureReactErrorBoundaryException` is because it can be used in other
scenarios as well, like with the new React 19 error APIs. Therefore it
was generalized to just be `captureReactException`.