profiling: add platform to chunk envelope (#15758)
We need to send the platform as part of envelope headers as this is the
item that relay is checking to manage rate limiting. We only care about
backend (node) and in the future frontend.

@AbhiPrasad please correct me here to use the correct convention, would
`javascript.node` be more appropriate? The only requirement really is to
have all nodejs platform types be of a single platform type, so we dont
need to maintain the ever growing list in relay