fix(nuxt): Delete no longer needed Nitro 'close' hook (#15790)
Due to PR https://github.com/getsentry/sentry-javascript/pull/15710 the
Sentry server config is only processed inside the Nitro-part of Nuxt and
the server config file is emitted through the Rollup plugin.

It is not needed anymore to copy-paste the file manually (this has been
the case before as the file was added to the `.nuxt` folder - as it was
processed by Nuxt, not Nitro-only).

This fixes the "no such file" error [posted in this
comment](https://github.com/getsentry/sentry-javascript/issues/13330#issuecomment-2743702460).

closes https://github.com/getsentry/sentry-javascript/issues/13330