feat(browser): Add `diagnoseSdkConnectivity()` function to programmatically detect possible connectivity issues (#15821)
Resolves https://github.com/getsentry/sentry-javascript/issues/15780

Adds a function `diagnoseSdkConnectivity` that will resolve to various
"error codes" for possible reasons why events might not land in Sentry.

Originally I wanted to add an API to detect ad blockers so we can show a
notification in the example pages that the wizard creates. Then I
quickly thought of a different case why the SDK might not send data so I
extended the API to also detect if the SDK wasn't initialized yet. We
can extend this function at any point if we come up with more cases.