feat(core): Add support for parameterizing logs (#15812)
ref https://github.com/getsentry/sentry-javascript/issues/15526

This adds support for parameterizing logs via the existing
`ParameterizedString` type and `parameterize` function exported from the
SDK. This works for all usages of the logger, so browser and Node.js.

Usage:

```js
Sentry.logger.info(Sentry.logger.fmt`User ${user} updated profile picture`, {
  userId: 'user-123',
  imageSize: '2.5MB',
  timestamp: Date.now()
});
```

`fmt` is an alias to `Sentry.parameterize` that is exported from the
`logger` namespace.

To support this change, I changed the typing of `ParameterizedString` to
accept `unknown[]` for `__sentry_template_values__`. This is broadening
the type, so should not be a breaking change.
[`logentry.params`](https://github.com/getsentry/relay/blob/a91f0c92860f88789ad6092ef5b1062aa3e34b80/relay-event-schema/src/protocol/logentry.rs#L51C27-L51C32)
should accept all kinds of values, relay handles formatting them
correctly.