ref(core): Avoid using `SentryError` for event processing control flow (#15823)
This PR replaces our usage of `SentryError` for control flow in our
event processing.
Instead, we now throw either an `InternalError` or `DoNotSendEventError`
(which are just POJOs with a symbol, without a stackframe). These two
also allow us to differentiate between these two use cases, which so far
have been kind of combined via the log level, but are really different
things - one is "expected"/configured by a user, the other is unexpected
and more of a warning.

I also removed the handling for `SentryError` from inbound filters, as
this should then become unused.

This also deprecates `SentryError`, it is no longer used and we can
eventually remove it.

ref
https://github.com/getsentry/sentry-javascript/issues/15725#issuecomment-2750892929