ref: Rewrite grpc test to be less accurate, but more stable
This rewrite makes the tests less accurate because we are no longer
testing against google-gax directly, but this feels like something we
should be doing in our integration tests instead.

As an alternative we check that we are filling the prototype correctly,
and then if we are patching the google-gax stub properly. This also
allows us to drop googleapis and google-gax as dev deps.