feat(node): Only add span listeners for instrumentation when used (#15802)
Related to
https://github.com/getsentry/sentry-javascript/issues/15725#issuecomment-2740015011,
this PR is an idea to avoid registering `on('spanStart')` hooks for all
the node libraries that do not have proper hooks, unless they are used.

Today, for OTEL instrumentation that does not provide span hooks, we
fall back to `client.on('spanStart')` to enhance/mutate spans emitted by
the instrumentation. This PR improved this by only registering the
`spanStart` client hook if we detect that the OTEL instrumentation is
actually wrapping something. This avoids us calling a bunch of span
callbacks each time a span is started, when it is not even needed.

For this, a new `instrumentWhenWrapped` utility is added which can be
passed an instrumentation. This works by monkey-patching `_wrap` on the
instrumentation, and ensuring a callback is only called conditionally.

Note: For some (e.g. connect, fastify) we register `spanStart` in the
error handler, which is even better, so there we do not need this logic.