ref(opentelemetry): Avoid sampling work for non-root spans (#15820)
Related to
https://github.com/getsentry/sentry-javascript/issues/15725#issuecomment-2742022649

When a span is not a root span, we do not need to do sampling work for
it. By reordering stuff in the `shouldSample` function, we should be
able to save a bunch of operations for the vast majority of (non-root)
spans.

I _think_ this should be just fine!