ref(browser/core): Move all log flushing logic into clients (#15831)
I experimented with a bunch of different approaches to register log
flushing implementations, but I found the cleanest and most performant
solution was to have log flushing logic live in clients.

This refactors the browser client to do that, and adds tests for all of
our flushing logic everywhere.

The bundle size increase is not ideal, but I recognize we have no
choice.

After this gets merged in, I'll add the console logging integration to
core. I'll also add some quick integrations for pino and winston.