feat(nuxt): Base decision on source maps upload only on Nuxt source map settings (#15859)
There is a total of 6 settings in Nuxt where you can disable/enable your
source maps. This is not only difficult to understand when writing and
maintaining the logic for this. It may also be hard to debug stuff for
users.


![image](https://github.com/user-attachments/assets/73ff315e-2000-4408-ba62-39bd10083378)

There is this discussion around Nuxt source maps:
https://github.com/getsentry/sentry-javascript/discussions/15028
And this issue:
https://github.com/getsentry/sentry-javascript/issues/15160

A problem with the current setup was not only the difficulty to
understand all of this but also it sometimes just overwrote stuff you
didn't want it to. Like in the default case of Nuxt, `sourcemap.server`
is `true`, but as `nitro.rollupConfig.output.sourcemap` is always
undefined, Sentry would overwrite this setting to `'hidden'`, even
though the source maps were enabled already.

---

The only two relevant options in Nuxt are the [root-level sourcemap
options](https://nuxt.com/docs/guide/going-further/debugging#sourcemaps).
Those settings will propagate to nitro, vite and rollup.

As we overwrite the source maps setting if the setting is undefined,
only the basic Nuxt source map settings (linked above) are taken into
account for that from now on.