feat(core): Add `consoleLoggingIntegration` for logs (#15955)
Adds `consoleLoggingIntegration` integration that takes calls to
`console.X` and flushes them as logs. For now this is an opt-in
integration. We can evaluate it's default status at a later point.

```js
import * as Sentry from '@sentry/browser'; // or any other supported sdk

Sentry.init({
  // send console.log, console.error, and console.warn calls as logs to Sentry
  integrations: [Sentry.consoleLoggingIntegration({ levels: ['log', 'error', 'warn'] })],
});
```

## Notes

In general we have inconsistencies with console instrumentation across
our different SDKs.

- In browser we have console instrumentation as part of the
`Breadcrumbs` integration.
- In node we have console instrumentation that generates breadcrumbs in
the `Console` integration
- In core we have console instrumentation, which generates errors and
messages via the `CaptureConsole` integration

For now because logs are experimental, making it a standalone
integration feels fine to me. As we get closer to GA status for logs
(and there is a distinct path for breadcrumbs) we should evaluate how
all of our console integrations are structured.