feat(browser): Add `onRequestSpanStart` hook to browser tracing integration (#15979)
This PR adds a hook to the browser tracing integration that allows to
hook into the creation of request spans (fetch & xhr) and set attributes
based on headers (can be extended with additional info in the future).

Primarily I was gonna use this to set an attribute/op on prefetch client
requests in Next.js to show them as `http.client.prefetch`.

---------

Co-authored-by: Lukas Stracke <lukas.stracke@sentry.io>