feat(nextjs): Add `captureRouterTransitionStart` hook for capturing navigations (#15981)
We relied on Next.js internals for navigation spans and that inevitably
broke so Next.js added a `onRouterTransitionStart` hook in
`instrumentation-client.ts`
(https://github.com/vercel/next.js/pull/77791) for us to use.

This PR exposes a handler called `captureRouterTransitionStart` for that
hook so that we can actually instrument navigations again.