fix(node): Ensure late init works with all integrations (#16016)
In cases where we tweak the options we pass to the instrumentation, our
code did not handle this very well. This could lead to certain options
missing when init is called again later, as we updated the config with
incomplete/wrong options.

To fix this properly, I added a new variant to `generateInstrumentOnce`
which accepts a class and an options transformer, which ensures that the
options are always correctly transformed, no matter if called the first
or second time.

(The code to make this overloaded `generateInstrumentOnce` function work
with TS was pretty tricky, but I think it is good now - type inferral
etc. works nicely now!)

Fixes https://github.com/getsentry/sentry-javascript/issues/16004