feat(core): Move console integration into core and add to cloudflare/vercel-edge (#16024)
resolves https://github.com/getsentry/sentry-javascript/issues/15439
resolves https://github.com/getsentry/sentry-javascript/issues/4532

resolves https://linear.app/getsentry/issue/JSC-192

Supercedes https://github.com/getsentry/sentry-javascript/pull/16021

Our console instrumentation was always inconsistent, but adding
breadcrumbs should not really be. This PR adds a unified console
instrumentation to `@sentry/core`, and makes the Node SDK use that. We
also add this to the `vercel-edge` and `cloudflare` SDKs.

I also left todo comments in the deno and browser SDKs for us to unify
into this single integration afterwards.

---------

Co-authored-by: Francesco Gringl-Novy <francesco.novy@sentry.io>