feat(bun): Support new `Bun.serve` APIs (#16035)
Supercedes https://github.com/getsentry/sentry-javascript/pull/15978

resolves https://github.com/getsentry/sentry-javascript/issues/15941
resolves https://github.com/getsentry/sentry-javascript/issues/15827
resolves https://github.com/getsentry/sentry-javascript/issues/15816

Bun recently updated their `Bun.serve` API with new functionality, which
unfortunately broke our existing instrumentation. This is detailed with
https://bun.sh/docs/api/http#bun-serve. Specifically, they added a new
routing API that looks like so:

```ts
Bun.serve({
  // `routes` requires Bun v1.2.3+
  routes: {
    // Dynamic routes
    "/users/:id": req => {
      return new Response(`Hello User ${req.params.id}!`);
    },

    // Per-HTTP method handlers
    "/api/posts": {
      GET: () => new Response("List posts"),
      POST: async req => {
        const body = await req.json();
        return Response.json({ created: true, ...body });
      },
    },

    // Wildcard route for all routes that start with "/api/" and aren't otherwise matched
    "/api/*": Response.json({ message: "Not found" }, { status: 404 }),

    // Redirect from /blog/hello to /blog/hello/world
    "/blog/hello": Response.redirect("/blog/hello/world"),

    // Serve a file by buffering it in memory
    "/favicon.ico": new Response(await Bun.file("./favicon.ico").bytes(), {
      headers: {
        "Content-Type": "image/x-icon",
      },
    }),
  },

  // (optional) fallback for unmatched routes:
  // Required if Bun's version < 1.2.3
  fetch(req) {
    return new Response("Not Found", { status: 404 });
  },
});
``` 

Because there are now dynamic routes and wildcard routes, we can
actually generate `route` transaction source and send parameterized
routes to Sentry. The `fetch` API is still supported.

The only API we don't support is [static
routes/responses](https://bun.sh/docs/api/http#static-responses). This
is because these are optimized by Bun itself, and if we turn it into a
function (which we need to do to time it), we'll lose out on the
optimization. For now they aren't instrumented.