test(replay): Update replay tests to handle side effects explicitly (#16135)
Extracted out of
https://github.com/getsentry/sentry-javascript/pull/16134

During that refactor, a bunch of stuff failed. The reason for this is
that we relied on a side-effect mock in `use-fake-timers` utility, which
is really not ideal at all.

So this refactors this to instead make this explicit, as a side-effect
import on the top of the test files. Then we can just use
`vi.useFakeTimers()` normally.