fix(remix): Avoid rewrapping root loader. (#16136)
Resolves: https://github.com/getsentry/sentry-javascript/issues/16075

We used to allow rewrapping of the root loader when we had a separate
Express adapter, where we could not determine the instrumentation order.
Now that we don't have it anymore, we can remove this. Still, can't
reproduce the original issue running e2e test apps under load. But it's
clear that's what caused it.