build: Actually use import eslint rules (#16134)
We had a bunch of import/export related eslint rules, but did not
actually apply them to anything 🤦 this PR changes this, and also applies
all the rules to all the files.

The vast majority of the changes is just auto-fixes, but there are a few
small manual things that were needed - these are in separate commits,
it's probably easier to look at them separately!

It also updates prettier to apply to dev-packages.