test(replay): Rewrite custom matcher to be recognized by VSCode (#16141)
I was always annoyed that this was not properly highlighted by VSCode,
so I looked into this a bit, and actually refactored it a bit to be more
type-safe in the vitest world we now live in :)

I got rid of `toHaveSentReplay` in favor of simply always using
`toHaveLastSentReplay` (which we already used almost everywhere). I also
moved the type definition into a separate file and ensured it actually
works, at leas tin my vscode/cursor instance. This also kind of surfaced
some type errors that existed in tests, but have not been shown (??) for
whatever reason... no idea what is going on there.