fix: Vercel `ai` ESM patching (#16152)
- Closes #16137 

With ESM patching you need to retain the original module and just
overwrite the exports you want to wrap. This usually works for CJS too
but unfortunately, the CJS exports of `ai` only have getters and no
setters so this route is not possible.

This PR changes the patching so that it works slightly differently for
ESM and CJS. The original code that outputs a newly created object is
retained for CJS but for ESM we use the preferred route of replacing the
required exports.

To detect whether the module we're patching is an ES module we check
`Object.prototype.toString.call(moduleExports) === '[object Module]'`
which is documented
[here](https://tc39.es/ecma262/#sec-module-namespace-objects).

This PR also adds an ESM test for `ai`.