test(node): Add utility to test esm & cjs instrumentation (#16159)
Today, we do not have a lot of esm-specific node integration tests. Our
tests make it possible to test ESM, but we only use it very rarely,
sticking to cjs tests across the suite mostly. This means that we have a
bunch of gaps around our ESM support.

This PR introduces a new test utility to make it easier to test stuff in
ESM and CJS:

```js
createEsmAndCjsTests(
    __dirname,
    'scenario.mjs',
    'instrument.mjs',
    (createRunner, test) => {
      test('it works when importing the http module', async () => {
        const runner = createRunner(); // normal test as before
       });
});
```

This has a few limitations based on how this works - we can make this
more robust in the future, but for now it should be "OK":
1. It requires a .mjs based instrument file as well as an .mjs based
scenario
2. No relative imports are supported (all content must be in these two
files)
3. It simply regex replaces the esm imports with require for the CJS
tests. not perfect, but it kind of works

For tests that are known to fail on e.g. esm or cjs, you can configure
`failsOnEsm: true`. In this case, it will fail if the test _does not
fail_ (by using `test.fails()` to ensure test failure). This way we can
ensure we find out if stuff starts to fail etc.

To make this work, I had to re-write the test runner code a bit, because
it had issues with vitest unhandled rejection handling. Due to the way
we handled test completion with a promise, `test.fails` was not working
as expected, because the test indeed succeeded when it failed, but the
overall test run still failed because an unhandled rejection bubbled up.
Now, this should work as expected...

I re-wrote a few tests already to show how it works, plus added a new
test that shows an ESM test failure when importing http module (😭 )