fix(feedback): Prevent `removeFromDom()` from throwing (#16030)
An internal user reported seeing "NotFoundError: Failed to execute
'removeChild' on 'Node': The node to be removed is not a child of this
node." and the replay of their session confirms it:
https://sentry.sentry.io/explore/replays/127444034ae84099a84b524458b6dd90/

However, there is no matching JS error to give more clues.

What I think happened is that after interacting with the Feedback SDK
the widget got into a state where it was no longer mounted into the
page, but we called `removeFromDom()` anyway.

Looking at the replay i saw the moment when the feedback dom was aded to
the page, but it wasn't visible, only this got added at 18:51:
<img width="265" alt="SCR-20250411-lsav"
src="https://github.com/user-attachments/assets/1f65a174-1a8a-405c-8ce4-be0a10d09a64"
/>

So something prevented it from opening all the way up.

Fixes https://github.com/getsentry/sentry/issues/89424