fix(node): Avoid double-wrapping http module (#16177)
@s1gr1d and @Lms24 figured out that the duplicate spans we sometimes see
are related to esm and the http module. Especially, it seems to be
related to us using the `stealthWrap` function to wrap `server.emit` for
request isolation purposes.

While we still don't really know _why_ this is making such problems,
this PR seems to fix it (at least in integration tests) by avoiding
using import-in-the-middle here, and instead using diagnostics channel
with good old-fashioned monkey patching on the passed-in `server`
instance.

Some note: We need to make sure to still call this in the otel-wrapping
code of `init()`, otherwise there are weird timing issues in
top-level-import scenarios 😬

Hopefully fixes
https://github.com/getsentry/sentry-javascript/issues/15830, and fixes
https://github.com/getsentry/sentry-javascript/issues/15803

---------

Co-authored-by: Andrei <168741329+andreiborza@users.noreply.github.com>