feat(browser): Add option to sample linked traces consistently (#16037)
Implement consistent sampling across traces in browser SDKs.

Concrete changes:
- Add `consistentTraceSampling` option
- Emit `beforeSampling` hook before sampling and creating `SentrySpan`'s
in the coreSDK (this hook is already called in Otel-based spans).
- Allow `beforeSampling` hook subscribers to alter span `attributes` and
`parentSampled` flag
- Subscribe to `beforeSampling` hook when `consistentTraceSampling` is
enabled and `linkPreviousTrace` is not disabled
- modify attributes and `parentSampled` flag based on the sampling
decision, sample rate and sample rand of the previous trace
- this will ensure that the newly started trace is sampled
postitively/negatively depending on the decision of the previous trace
- we also directly modify the propagation context to ensure that the
previous trace's sampling decision, rate and rand are propagated
correctly to downstream services and in the `trace` envelope header.
This is necessary to ensure correct span metrics extrapolation.
- sampling decisions from `<meta>` tags still have precedence for the
first pageload over the previous trace's sampling decision. This is
necessary because we need to prioritize inter-trace completeness over
intra-trace completeness
- explicit sampling decisions from `tracesSampler` still have precedence
over the previous trace's sampling decision
- added a bunch of browser integration tests describing the behaviour in
various edge cases and sampling configurations