feat(cloudflare): Add support for email, queue, and tail handler (#16233)
Cloudflare workers have a variety of handlers you can define for
different functionality:
https://developers.cloudflare.com/workers/runtime-apis/handlers/

Right now our instrumentation wraps the `fetch` and `scheduled`
handlers. This PR extends our instrumentation to also wrap the `queue`,
`email`, and `tail` handler.

We only create spans for queue and email, but not for tail because its
meant as a debugging/analytics endpoint. We can introduce spans in
`tail` handler if there is demand in the future.

We need to add this wrapping, otherwise users cannot use the SDK in
these handlers because of how the cloudflare isolation model works
(fetch handler is completed isolated from email and queue handlers, so
we need to wrap every handler individually).